Physio Notes/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
anthropic
python-dotenv
aiofiles
